Refactor model: Move "Connection" in separate class file.
